requirements.txt
Flask==3.0.0
openai-whisper
torch
werkzeug==3.0.1
numpy<2